move common structures from barretenberg_rs to barretenberg_structures.
We do this because the serialiser needs these and barretenberg_rs is not always available due to the feature flags
