+++ src/barretenberg_rs/composer.rs
@@ -1,6 +1,6 @@
 use super::crs::CRS;
 use super::pippenger::Pippenger;
-use acvm::FieldElement as Scalar;
+use crate::barretenberg_structures::*;
 use std::slice;
 
 pub struct StandardComposer {
@@ -22,465 +22,6 @@ impl StandardComposer {
             crs,
             constraint_system,
         }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Assignments(pub(crate) Vec<Scalar>);
-pub type WitnessAssignments = Assignments;
-
-impl Assignments {
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let witness_len = self.0.len() as u32;
-        buffer.extend_from_slice(&witness_len.to_be_bytes());
-
-        for assignment in self.0.iter() {
-            buffer.extend_from_slice(&assignment.to_bytes());
-        }
-
-        buffer
-    }
-
-    pub fn from_vec(vec: Vec<Scalar>) -> Assignments {
-        Assignments(vec)
-    }
-
-    pub fn push_i32(&mut self, value: i32) {
-        self.0.push(Scalar::from(value as i128));
-    }
-    pub fn push(&mut self, value: Scalar) {
-        self.0.push(value);
-    }
-    pub fn new() -> Assignments {
-        Assignments(vec![])
-    }
-}
-
-impl Default for Assignments {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Clone, Hash, Debug)]
-pub struct Constraint {
-    pub a: i32,
-    pub b: i32,
-    pub c: i32,
-    pub qm: Scalar,
-    pub ql: Scalar,
-    pub qr: Scalar,
-    pub qo: Scalar,
-    pub qc: Scalar,
-}
-
-impl Constraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-        // serialise Wires
-        buffer.extend_from_slice(&self.a.to_be_bytes());
-        buffer.extend_from_slice(&self.b.to_be_bytes());
-        buffer.extend_from_slice(&self.c.to_be_bytes());
-
-        // serialise selectors
-        buffer.extend_from_slice(&self.qm.to_bytes());
-        buffer.extend_from_slice(&self.ql.to_bytes());
-        buffer.extend_from_slice(&self.qr.to_bytes());
-        buffer.extend_from_slice(&self.qo.to_bytes());
-        buffer.extend_from_slice(&self.qc.to_bytes());
-
-        buffer
-    }
-}
-
-#[derive(Clone, Hash, Debug)]
-pub struct RangeConstraint {
-    pub a: i32,
-    pub num_bits: i32,
-}
-
-impl RangeConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-        // Serialiasing Wires
-        buffer.extend_from_slice(&self.a.to_be_bytes());
-        buffer.extend_from_slice(&self.num_bits.to_be_bytes());
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct EcdsaConstraint {
-    pub hashed_message: Vec<i32>,
-    pub signature: [i32; 64],
-    pub public_key_x: [i32; 32],
-    pub public_key_y: [i32; 32],
-    pub result: i32,
-}
-
-impl EcdsaConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let message_len = (self.hashed_message.len()) as u32;
-        buffer.extend_from_slice(&message_len.to_be_bytes());
-        for constraint in self.hashed_message.iter() {
-            buffer.extend_from_slice(&constraint.to_be_bytes());
-        }
-
-        let sig_len = (self.signature.len()) as u32;
-        buffer.extend_from_slice(&sig_len.to_be_bytes());
-        for sig_byte in self.signature.iter() {
-            buffer.extend_from_slice(&sig_byte.to_be_bytes());
-        }
-
-        let pub_key_x_len = (self.public_key_x.len()) as u32;
-        buffer.extend_from_slice(&pub_key_x_len.to_be_bytes());
-        for x_byte in self.public_key_x.iter() {
-            buffer.extend_from_slice(&x_byte.to_be_bytes());
-        }
-
-        let pub_key_y_len = (self.public_key_y.len()) as u32;
-        buffer.extend_from_slice(&pub_key_y_len.to_be_bytes());
-        for y_byte in self.public_key_y.iter() {
-            buffer.extend_from_slice(&y_byte.to_be_bytes());
-        }
-
-        buffer.extend_from_slice(&self.result.to_be_bytes());
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct SchnorrConstraint {
-    pub message: Vec<i32>,
-    pub signature: [i32; 64],
-    pub public_key_x: i32,
-    pub public_key_y: i32,
-    pub result: i32,
-}
-
-impl SchnorrConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let message_len = (self.message.len()) as u32;
-        buffer.extend_from_slice(&message_len.to_be_bytes());
-        for constraint in self.message.iter() {
-            buffer.extend_from_slice(&constraint.to_be_bytes());
-        }
-
-        let sig_len = (self.signature.len()) as u32;
-        buffer.extend_from_slice(&sig_len.to_be_bytes());
-        for sig_byte in self.signature.iter() {
-            buffer.extend_from_slice(&sig_byte.to_be_bytes());
-        }
-
-        buffer.extend_from_slice(&self.public_key_x.to_be_bytes());
-        buffer.extend_from_slice(&self.public_key_y.to_be_bytes());
-        buffer.extend_from_slice(&self.result.to_be_bytes());
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct MerkleMembershipConstraint {
-    pub hash_path: Vec<i32>,
-    pub root: i32,
-    pub leaf: i32,
-    pub index: i32,
-    pub result: i32,
-}
-
-impl MerkleMembershipConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let hash_path_len = self.hash_path.len() as u32;
-
-        buffer.extend_from_slice(&hash_path_len.to_be_bytes());
-        for constraint in self.hash_path.iter() {
-            buffer.extend_from_slice(&constraint.to_be_bytes());
-        }
-
-        buffer.extend_from_slice(&self.root.to_be_bytes());
-        buffer.extend_from_slice(&self.leaf.to_be_bytes());
-        buffer.extend_from_slice(&self.result.to_be_bytes());
-        buffer.extend_from_slice(&self.index.to_be_bytes());
-
-        buffer
-    }
-}
-
-#[derive(Clone, Hash, Debug)]
-pub struct Sha256Constraint {
-    pub inputs: Vec<(i32, i32)>,
-    pub result: [i32; 32],
-}
-
-impl Sha256Constraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let inputs_len = self.inputs.len() as u32;
-        buffer.extend_from_slice(&inputs_len.to_be_bytes());
-        for constraint in self.inputs.iter() {
-            buffer.extend_from_slice(&constraint.0.to_be_bytes());
-            buffer.extend_from_slice(&constraint.1.to_be_bytes());
-        }
-
-        let result_len = self.result.len() as u32;
-        buffer.extend_from_slice(&result_len.to_be_bytes());
-        for constraint in self.result.iter() {
-            buffer.extend_from_slice(&constraint.to_be_bytes());
-        }
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct Blake2sConstraint {
-    pub inputs: Vec<(i32, i32)>,
-    pub result: [i32; 32],
-}
-
-impl Blake2sConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let inputs_len = self.inputs.len() as u32;
-        buffer.extend_from_slice(&inputs_len.to_be_bytes());
-        for constraint in self.inputs.iter() {
-            buffer.extend_from_slice(&constraint.0.to_be_bytes());
-            buffer.extend_from_slice(&constraint.1.to_be_bytes());
-        }
-
-        let result_len = self.result.len() as u32;
-        buffer.extend_from_slice(&result_len.to_be_bytes());
-        for constraint in self.result.iter() {
-            buffer.extend_from_slice(&constraint.to_be_bytes());
-        }
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct HashToFieldConstraint {
-    pub inputs: Vec<(i32, i32)>,
-    pub result: i32,
-}
-
-impl HashToFieldConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let inputs_len = self.inputs.len() as u32;
-        buffer.extend_from_slice(&inputs_len.to_be_bytes());
-        for constraint in self.inputs.iter() {
-            buffer.extend_from_slice(&constraint.0.to_be_bytes());
-            buffer.extend_from_slice(&constraint.1.to_be_bytes());
-        }
-
-        buffer.extend_from_slice(&self.result.to_be_bytes());
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct PedersenConstraint {
-    pub inputs: Vec<i32>,
-    pub result_x: i32,
-    pub result_y: i32,
-}
-
-impl PedersenConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-
-        let inputs_len = self.inputs.len() as u32;
-        buffer.extend_from_slice(&inputs_len.to_be_bytes());
-        for constraint in self.inputs.iter() {
-            buffer.extend_from_slice(&constraint.to_be_bytes());
-        }
-
-        buffer.extend_from_slice(&self.result_x.to_be_bytes());
-        buffer.extend_from_slice(&self.result_y.to_be_bytes());
-
-        buffer
-    }
-}
-#[derive(Clone, Hash, Debug)]
-pub struct FixedBaseScalarMulConstraint {
-    pub scalar: i32,
-    pub pubkey_x: i32,
-    pub pubkey_y: i32,
-}
-
-impl FixedBaseScalarMulConstraint {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-        // Serialising Wires
-        buffer.extend_from_slice(&self.scalar.to_be_bytes());
-        buffer.extend_from_slice(&self.pubkey_x.to_be_bytes());
-        buffer.extend_from_slice(&self.pubkey_y.to_be_bytes());
-
-        buffer
-    }
-}
-
-#[derive(Clone, Hash, Debug)]
-pub struct LogicConstraint {
-    a: i32,
-    b: i32,
-    result: i32,
-    num_bits: i32,
-    is_xor_gate: bool,
-}
-
-impl LogicConstraint {
-    pub fn and(a: i32, b: i32, result: i32, num_bits: i32) -> LogicConstraint {
-        LogicConstraint {
-            a,
-            b,
-            result,
-            num_bits,
-            is_xor_gate: false,
-        }
-    }
-    pub fn xor(a: i32, b: i32, result: i32, num_bits: i32) -> LogicConstraint {
-        LogicConstraint {
-            a,
-            b,
-            result,
-            num_bits,
-            is_xor_gate: true,
-        }
-    }
-
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer = Vec::new();
-        // Serialising Wires
-        buffer.extend_from_slice(&self.a.to_be_bytes());
-        buffer.extend_from_slice(&self.b.to_be_bytes());
-        buffer.extend_from_slice(&self.result.to_be_bytes());
-        buffer.extend_from_slice(&self.num_bits.to_be_bytes());
-        buffer.extend_from_slice(&i32::to_be_bytes(self.is_xor_gate as i32));
-
-        buffer
-    }
-}
-
-#[derive(Clone, Hash, Debug)]
-pub struct ConstraintSystem {
-    pub var_num: u32,
-    pub public_inputs: Vec<u32>,
-
-    pub logic_constraints: Vec<LogicConstraint>,
-    pub range_constraints: Vec<RangeConstraint>,
-    pub sha256_constraints: Vec<Sha256Constraint>,
-    pub merkle_membership_constraints: Vec<MerkleMembershipConstraint>,
-    pub schnorr_constraints: Vec<SchnorrConstraint>,
-    pub ecdsa_secp256k1_constraints: Vec<EcdsaConstraint>,
-    pub blake2s_constraints: Vec<Blake2sConstraint>,
-    pub pedersen_constraints: Vec<PedersenConstraint>,
-    pub hash_to_field_constraints: Vec<HashToFieldConstraint>,
-    pub fixed_base_scalar_mul_constraints: Vec<FixedBaseScalarMulConstraint>,
-    pub constraints: Vec<Constraint>,
-}
-
-impl ConstraintSystem {
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut buffer: Vec<u8> = Vec::new();
-
-        // Push lengths onto the buffer
-        buffer.extend_from_slice(&self.var_num.to_be_bytes());
-
-        let pi_len = self.public_inputs.len() as u32;
-        buffer.extend_from_slice(&pi_len.to_be_bytes());
-        for pub_input in self.public_inputs.iter() {
-            buffer.extend_from_slice(&pub_input.to_be_bytes());
-        }
-
-        // Serialise each Logic constraint
-        let logic_constraints_len = self.logic_constraints.len() as u32;
-        buffer.extend_from_slice(&logic_constraints_len.to_be_bytes());
-        for constraint in self.logic_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Range constraint
-        let range_constraints_len = self.range_constraints.len() as u32;
-        buffer.extend_from_slice(&range_constraints_len.to_be_bytes());
-        for constraint in self.range_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Sha256 constraint
-        let sha256_constraints_len = self.sha256_constraints.len() as u32;
-        buffer.extend_from_slice(&sha256_constraints_len.to_be_bytes());
-        for constraint in self.sha256_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Merkle Membership constraint
-        let merkle_membership_constraints_len = self.merkle_membership_constraints.len() as u32;
-        buffer.extend_from_slice(&merkle_membership_constraints_len.to_be_bytes());
-        for constraint in self.merkle_membership_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Schnorr constraint
-        let schnorr_len = self.schnorr_constraints.len() as u32;
-        buffer.extend_from_slice(&schnorr_len.to_be_bytes());
-        for constraint in self.schnorr_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each ECDSA constraint
-        let ecdsa_len = self.ecdsa_secp256k1_constraints.len() as u32;
-        buffer.extend_from_slice(&ecdsa_len.to_be_bytes());
-        for constraint in self.ecdsa_secp256k1_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Blake2s constraint
-        let blake2s_len = self.blake2s_constraints.len() as u32;
-        buffer.extend_from_slice(&blake2s_len.to_be_bytes());
-        for constraint in self.blake2s_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Pedersen constraint
-        let pedersen_len = self.pedersen_constraints.len() as u32;
-        buffer.extend_from_slice(&pedersen_len.to_be_bytes());
-        for constraint in self.pedersen_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each HashToField constraint
-        let h2f_len = self.hash_to_field_constraints.len() as u32;
-        buffer.extend_from_slice(&h2f_len.to_be_bytes());
-        for constraint in self.hash_to_field_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each HashToField constraint
-        let fixed_base_scalar_mul_len = self.fixed_base_scalar_mul_constraints.len() as u32;
-        buffer.extend_from_slice(&fixed_base_scalar_mul_len.to_be_bytes());
-        for constraint in self.fixed_base_scalar_mul_constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        // Serialise each Arithmetic constraint
-        let constraints_len = self.constraints.len() as u32;
-        buffer.extend_from_slice(&constraints_len.to_be_bytes());
-        for constraint in self.constraints.iter() {
-            buffer.extend(&constraint.to_bytes());
-        }
-
-        buffer
     }
 }
