+++ src/lib.rs
@@ -8,3 +8,4 @@ pub mod serialiser;
 pub use serialiser::serialise_circuit;
 pub mod acvm_interop;
 pub use acvm_interop::Plonk;
+mod barretenberg_structures;
